requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.45.0
django-debug-toolbar==3.2
django-filter==2.4.0
idna==3.2
oauthlib==3.1.1
Pillow==8.1.2
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.6
